First commit
A Pokémon type calculator, accurate for Gen VI. It has a basic and
advanced mode. Basic mode allows for selecting attack type and defending
type(s), and tells you how effective the move will be. Advanced mode
allows you to set attacking and defending stats, as well as the power
and type of the move and the types of the attacking Pokémon, which will
then run calculations and determine how much damage the attack will do.